🎨 style(i18nConfig.ts): add const assertion to i18nConfig object 🆕 feat(i18nConfig.ts): add Locale type The const assertion to the i18nConfig object ensures that the object is not accidentally modified. The Locale type is added to ensure that the locale parameter is one of the supported locales.
🆕 feat: add en.json dictionary
🆕 feat: add pt.json dictionary
The en.json and pt.json dictionaries are added to support the English and Portuguese languages.

🆕 feat(getDictionary.ts): add getDictionary function
The getDictionary function is added to retrieve the dictionary for a given locale.

📦 chore(package.json): remove accept-language-parser and add negotiator
The accept-language-parser package is removed, and the negotiator package is added as a replacement.

🚚 chore(Navbar): move Navbar component to its own folder
The Navbar component is moved to its own

🐛 fix(middleware.ts): use Negotiator to parse Accept-Language header
✨ feat(middleware.ts): add support for language negotiation
The middleware now uses the Negotiator library to parse the Accept-Language header, which simplifies the code and improves performance. The middleware now supports language negotiation, which means that if the user's preferred language is not available, the middleware will try to find the best match. Additionally, the middleware now removes the language prefix from the URL if it is not needed.

✨ feat(tsconfig.json): add path mapping for dictionaries
The tsconfig.json file now includes a path mapping for the dictionaries folder, which allows for easier importing of dictionary files.