✨ feat(toaster.tsx): add toast component and provider The Toast component is a customizable toast notification component that can be used to display messages to the user. The Toaster component is a provider that manages the state of the toast notifications and renders them. It uses the Toast component to render each individual toast notification. The useToast hook is used to add new toast notifications to the Toaster component.
🚀 feat(ui): add use-toast hook and ToastMessageContent interface
The use-toast hook is inspired by the react-hot-toast library and provides a way to display toast messages. The hook returns a toast function that can be called with a ToastMessageContent object to display a toast message. The ToastMessageContent interface is also added to the dictionaries file to provide a way to define the content of the toast messages. The use-toast hook and ToastMessageContent interface improve the user experience by providing a way to display important information to the user in a non-intrusive way.