🔄 chore(next.config.js): update remotePatterns to include Shopify CDN 🔄 chore(route.ts): rename getProduct to getShopifyProduct and update import 🔄 chore(route.ts): rename getProducts to getShopifyProducts and update import 🔄 chore(layout.tsx): rename getCategories and getProducts to getShopifyCategories and getShopifyProducts, respectively. Update imports and add convertToShopifyLanguage function 🔄 chore(page.tsx): rename getProduct to getShopifyProduct and update import. Update function parameter to include handle and language 🔄 chore(page.tsx): rename getProducts to getShopifyProducts and update import. Update function parameter to include language 🔄 chore(page.tsx): update generateStaticParams function to use handle instead of id 🔄 chore(products/page.tsx): rename getCategories and getProducts to getShopifyCategories and getShopifyProducts, respectively. Update imports and add convertToShopifyLanguage function
🔥 chore(ENProducts.ts): remove unused file
The file ENProducts.ts was removed as it was no longer being used in the application.

🗑️ chore(PTProducts.ts): remove unused code
The file PTProducts.ts contained unused code that was removed to improve code readability and maintainability.

🔥 chore(products): remove unused files and interfaces
The files ProductsImages.ts, fetchProducts.tsx, and products.d.ts were removed as they were no longer being used in the project. The Product interface was also removed as it was being replaced by the new Product interface in the shopify/types.ts file.

✨ feat(shopify): add support for product categories and reshapeProduct function
The reshapeProduct function was updated to include the sizes, colors, and category properties of a product. The getShopifyCategories function was added to fetch all the categories of the products in the Shopify store. The reshapeCategories function was added to reshape the categories of the products into an array of strings.

🔀 merge(types): update import paths for Product type
The import paths for the Product type have been updated to reflect the new location of the type in the shopify/types file. This change was made to improve the organization of the codebase and to make it easier to find the type definition.